# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoxApp API
  description: |
    High-performance concurrent HTTP load testing tool with DNS timing metrics.
    
    This API provides endpoints for:
    - **Outgoing Traffic Management**: Configure and control external API calls
    - **Incoming Traffic Simulation**: Define mock routes with configurable response patterns
    - **Metrics & Monitoring**: Real-time metrics collection and health monitoring
    - **Scheduler Control**: Pause, resume, or emergency stop the load test
  version: 1.0.0
  contact:
    name: MoxApp Team
  license:
    name: Proprietary
    identifier: LicenseRef-MoxApp
servers:
  - url: __SERVER_URL__
    description: Current server
...
paths:
...
  /api/outgoing/settings/multiplier:
    get:
      tags:
        - Outgoing Settings
      summary: Get outgoing load multiplier
      description: Returns the current global load multiplier for outgoing requests
      operationId: getOutgoingMultiplier
      responses:
        '200':
          description: Current multiplier value
          content:
            application/json:
              schema:
                type: object
                properties:
                  global_multiplier:
                    type: number
                    format: double
                    example: 1.0
    post:
      tags:
        - Outgoing Settings
      summary: Set outgoing load multiplier
      description: Updates the global load multiplier for outgoing requests. Use 0.5 for 50% load, 2.0 for 200% load, etc.
      operationId: setOutgoingMultiplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - multiplier
              properties:
                multiplier:
                  type: number
                  format: double
                  minimum: 0
                  description: New multiplier value (must be non-negative)
                  example: 1.5
      responses:
        '200':
          description: Multiplier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplierUpdateResponse'
        '400':
          description: Invalid multiplier value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/settings/concurrency:
    get:
      tags:
        - Outgoing Settings
      summary: Get concurrent requests limit
      description: Returns the current maximum number of concurrent outgoing requests
      operationId: getOutgoingConcurrency
      responses:
        '200':
          description: Current concurrency limit
          content:
            application/json:
              schema:
                type: object
                properties:
                  concurrent_requests:
                    type: integer
                    example: 30
    post:
      tags:
        - Outgoing Settings
      summary: Set concurrent requests limit
      description: Updates the maximum number of concurrent outgoing requests allowed
      operationId: setOutgoingConcurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - concurrent
              properties:
                concurrent:
                  type: integer
                  minimum: 1
                  description: New concurrency limit (must be positive)
                  example: 50
      responses:
        '200':
          description: Concurrency limit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConcurrencyUpdateResponse'
        '400':
          description: Invalid concurrency value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/settings/log-requests:
    get:
      tags:
        - Outgoing Settings
      summary: Get outgoing request logging setting
      description: Returns whether individual outgoing requests are being logged
      operationId: getOutgoingLogRequests
      responses:
        '200':
          description: Current log requests setting
          content:
            application/json:
              schema:
                type: object
                properties:
                  log_all_requests:
                    type: boolean
                    example: false
    post:
      tags:
        - Outgoing Settings
      summary: Set outgoing request logging
      description: Enables or disables logging of individual outgoing requests
      operationId: setOutgoingLogRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - log_requests
              properties:
                log_requests:
                  type: boolean
                  description: Whether to log all individual requests
                  example: true
      responses:
        '200':
          description: Log requests setting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogRequestsUpdateResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/endpoints:
    get:
      tags:
        - Outgoing Endpoints
      summary: List all outgoing endpoints
      description: Returns all configured outgoing endpoints
      operationId: listOutgoingEndpoints
      responses:
        '200':
          description: List of endpoints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointListResponse'
    post:
      tags:
        - Outgoing Endpoints
      summary: Create outgoing endpoint
      description: Creates a new outgoing endpoint configuration
      operationId: createOutgoingEndpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointRequest'
      responses:
        '201':
          description: Endpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Endpoint already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/endpoints/{name}:
    get:
      tags:
        - Outgoing Endpoints
      summary: Get outgoing endpoint by name
      description: Returns a specific outgoing endpoint configuration
      operationId: getOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      responses:
        '200':
          description: Endpoint configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endpoint'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Outgoing Endpoints
      summary: Update outgoing endpoint
      description: Updates an existing outgoing endpoint configuration
      operationId: updateOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointRequest'
      responses:
        '200':
          description: Endpoint updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointUpdateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Outgoing Endpoints
      summary: Delete outgoing endpoint
      description: Deletes an outgoing endpoint by name
      operationId: deleteOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      responses:
        '200':
          description: Endpoint deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/endpoints/bulk:
    post:
      tags:
        - Outgoing Endpoints
      summary: Bulk create outgoing endpoints
      description: Creates multiple outgoing endpoints at once
      operationId: bulkCreateOutgoingEndpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/EndpointRequest'
      responses:
        '200':
          description: All endpoints created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResponse'
        '206':
          description: Partial success - some endpoints failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResponse'
        '400':
          description: All endpoints failed to create
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Outgoing Endpoints
      summary: Bulk delete outgoing endpoints
      description: Deletes multiple outgoing endpoints by name
      operationId: bulkDeleteOutgoingEndpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - names
              properties:
                names:
                  type: array
                  items:
                    type: string
                  description: List of endpoint names to delete
      responses:
        '200':
          description: All endpoints deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteResponse'
        '206':
          description: Partial success - some endpoints failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteResponse'
        '400':
          description: All endpoints failed to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/control:
    get:
      tags:
        - Outgoing Control
      summary: Get outgoing scheduler status
      description: Returns current outgoing scheduler control status
      operationId: getOutgoingControlStatus
      responses:
        '200':
          description: Scheduler status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlStatus'
        '503':
          description: Scheduler not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Outgoing Control
      summary: Control outgoing scheduler
      description: |
        Execute outgoing scheduler control actions:
        - `pause`: Stop scheduling new requests (in-flight requests complete)
        - `resume`: Resume scheduling requests
        - `emergency_stop`: Immediately stop all operations and cancel in-flight requests
      operationId: controlOutgoingScheduler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum:
                    - pause
                    - resume
                    - emergency_stop
                  description: Control action to execute
      responses:
        '200':
          description: Action executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlActionResponse'
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Scheduler not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/control/endpoint:
    post:
      tags:
        - Outgoing Control
      summary: Enable/disable outgoing endpoint
      description: Enables or disables a specific outgoing endpoint
      operationId: setOutgoingEndpointEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - enabled
              properties:
                name:
                  type: string
                  description: Endpoint name
                enabled:
                  type: boolean
                  description: Enable or disable the endpoint
      responses:
        '200':
          description: Endpoint status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointEnableResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/control/endpoints/bulk:
    post:
      tags:
        - Outgoing Control
      summary: Bulk enable/disable outgoing endpoints
      description: Enables or disables multiple outgoing endpoints at once
      operationId: bulkSetOutgoingEndpointsEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - names
                - enabled
              properties:
                names:
                  type: array
                  items:
                    type: string
                  description: List of endpoint names
                enabled:
                  type: boolean
                  description: Enable or disable the endpoints
      responses:
        '200':
          description: Endpoints status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkEnableResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/control/endpoints/all:
    post:
      tags:
        - Outgoing Control
      summary: Enable/disable all outgoing endpoints
      description: Enables or disables all outgoing endpoints
      operationId: setAllOutgoingEndpointsEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - enabled
              properties:
                enabled:
                  type: boolean
                  description: Enable or disable all endpoints
      responses:
        '200':
          description: All endpoints status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllEndpointsEnableResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/outgoing/auth-configs/{name}/status:
    get:
      tags:
        - Auth Configs
      summary: Get auth token status
      description: Returns token status for an auth config
      operationId: getAuthTokenStatus
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenStatus'
        '404':
          description: Auth config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Token manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/incoming/routes/reload:
    post:
      tags:
        - Incoming Routes
      summary: Reload routes from config
      description: |
        Reloads all incoming routes from the static configuration file.
        **Warning**: This replaces all runtime-created routes with the static configuration.
      operationId: reloadIncomingRoutes
      responses:
        '200':
          description: Routes reloaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteReloadResponse'
        '400':
          description: No config file path configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to reload config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MultiplierUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Global multiplier updated
        old_multiplier:
          type: number
          format: double
          example: 1.0
        new_multiplier:
          type: number
          format: double
          example: 1.5
    ConcurrencyUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Concurrent requests limit updated
        old_concurrent:
          type: integer
          example: 30
        new_concurrent:
          type: integer
          example: 50
    LogRequestsUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Log all requests setting updated
        old_log_requests:
          type: boolean
          example: false
        new_log_requests:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...